+++ mach2/mach.h
@@ -1,6 +1,6 @@
 #ifndef MACH_H
 #define MACH_H
 
-double mach(double, int);
+double mach(double, unsigned long int);
 
 #endif // MACH_H
