+++ reduc/zeta.h
@@ -1,6 +1,6 @@
 #ifndef ZETA_H
 #define ZETA_H
 
-double zeta(int, int);
+double zeta(unsigned long int, unsigned long int);
 
 #endif // ZETA_H
