+++ zeta1/zeta.c
@@ -1,6 +1,6 @@
-double zeta(int l, int u) {
+double zeta(unsigned long int l, unsigned long int u) {
 	double sum = 0;
-	for (int i = l; i <= u; i++) {
+	for (unsigned long int i = l; i <= u; i++) {
 		sum += 1.0/(i*i);
 	}
