+++ zeta2/num_pi.c
@@ -6,12 +6,12 @@
 int main(int argc, char **argv) {
 	if (argc != 2) return 1;
 
-	int n = atoi(argv[1]);
+	unsigned long int n = atoi(argv[1]);
 
 	double sum = zeta(n);
 
 	double num_pi = sqrt(6*sum);
-	printf("Accuracy: %f\n", fabs(M_PI - num_pi));
+	printf("Accuracy: %.16g\n", fabs(M_PI - num_pi));
 
 	return 0;
 }
