+++ zeta2/zeta.c
@@ -1,9 +1,9 @@
 #include <omp.h>
 
-double zeta(int n) {
+double zeta(unsigned long int n) {
 	double sum = 0;
 	#pragma omp parallel for schedule(static)
-	for (int i = 1; i <= n; i++) {
+	for (unsigned long int i = 1; i <= n; i++) {
 		sum += 1.0/(i*i);
 	}
